change --ignore-positions to include/exclude ranges
This replaces the argument `--ignore-positions`, which only accepts specific integer positions, with `--exclude-positions` which accepts numeric ranges or specific positions (ex. '100-200' or '100'). A related argument, `--include-positions`, is provided to allow the user to pass a restricted range or list of positions to be included (only). Position exclusion is considered after position inclusion. This also moves the x-axis scale labels to be adjacent to the genome map if `--flip-vertical` is specified.